Move position to separate file
