+++ yaml/shared/src/main/scala/org/virtuslab/yaml/internal/load/compose/Composer.scala
@@ -6,7 +6,7 @@ import org.virtuslab.yaml.YamlError
 import org.virtuslab.yaml.internal.load.parse.Event
 import org.virtuslab.yaml.internal.load.parse.ParserImpl
 import org.virtuslab.yaml.internal.load.reader.Scanner
-import org.virtuslab.yaml.internal.load.reader.Position
+import org.virtuslab.yaml.Position
 
 import scala.annotation.tailrec
