+++ yaml/shared/src/main/scala/org/virtuslab/yaml/internal/load/parse/Event.scala
@@ -1,7 +1,7 @@
 package org.virtuslab.yaml.internal.load.parse
 
 import org.virtuslab.yaml.internal.load.reader.token.ScalarStyle
-import org.virtuslab.yaml.internal.load.reader.Position
+import org.virtuslab.yaml.Position
 import org.virtuslab.yaml.internal.load.parse.Event.StreamStart
 import org.virtuslab.yaml.internal.load.reader.token.Token.SequenceStart
