+++ yaml/shared/src/main/scala/org/virtuslab/yaml/internal/load/reader/Reader.scala
@@ -3,13 +3,7 @@ package org.virtuslab.yaml.internal.load.reader
 import scala.util.Try
 import scala.annotation.tailrec
 
-final case class Position(offset: Int, line: Int, column: Int, input: Vector[String]):
-  def errorMsg: String =
-    val msg          = input(line - 1)
-    val spaces       = column - 1
-    val circumflexes = msg.length - spaces
-    s"""|$msg
-        |${" " * spaces}${"^" * circumflexes}""".stripMargin
+import org.virtuslab.yaml.Position
 
 trait Reader:
   def read(): Char
